fix monthly pipeline
